orders/admin.py
from django.contrib import admin
from django.utils.safestring import mark_safe
from unfold.admin import ModelAdmin, StackedInline, TabularInline

from .models import Order, OrderItem


# Register your models here.
class OrderItemInline(TabularInline):
    model = OrderItem
    raw_id_fields = ["product"]
    extra = 0


def order_payment(obj):
    url = obj.get_stripe_url()
    if obj.stripe_id:
        html = f'<a href="{url}" target="_blank">{obj.stripe_id}</a>'
        return mark_safe(html)
    return ""


order_payment.short_description = "Stripe Payment"


@admin.register(Order)
class OrderAdmin(ModelAdmin):
    list_display = [
        "id",
        "first_name",
        "last_name",
        "email",
        "address",
        "postal_code",
        "city",
        "paid",
        order_payment,
        "created",
        "updated",
    ]

    list_filter = ["paid", "created", "updated"]
    inlines = [OrderItemInline]
