pyproject.toml
[project]
name = "online-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-unfold>=0.73.1",
    "djlint>=1.36.4",
    "flower>=2.0.1",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "stripe>=14.0.1",
]
